Move guessing ocation country by provider to default location provider
